requirements.txt
git+https://github.com/geoffroypeeters/ssmnet_ISMIR2023.git
git+https://github.com/EleutherAI/aria.git
evaluate==0.4.1
jsonlines==4.0.0
miditoolkit==1.0.1
numpy
pandas==2.2.2
PyYAML==6.0.1
torch==2.1.0
tqdm==4.66.2
transformers==4.40.1
pretty_midi